image_processing/debug_regions/requirements.txt
pika
Pillow
jsonpath-rw
dpath
